app/requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
anthropic>=0.40.0
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0.0
